refactor: Product, MenuGroup 패키지 중심 의존성 분리 리팩토링
Product 에서 불필요한 로직 제거
MenuGroup은 별도로 의존중인 엔티티가 없음
