+++ src/main/java/kitchenpos/menugroup/application/MenuGroupService.java
@@ -2,7 +2,7 @@ package kitchenpos.menugroup.application;
 
 import java.util.List;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
+import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
 import kitchenpos.menugroup.ui.request.MenuGroupRequest;
 import kitchenpos.menugroup.ui.response.MenuGroupResponse;
 import org.springframework.stereotype.Service;
