+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -14,9 +14,9 @@ import kitchenpos.menu.ui.request.MenuRequest;
 import kitchenpos.menu.ui.response.MenuProductResponse;
 import kitchenpos.menu.ui.response.MenuResponse;
 import kitchenpos.menugroup.domain.MenuGroup;
-import kitchenpos.menugroup.repository.MenuGroupRepository;
+import kitchenpos.menugroup.domain.repository.MenuGroupRepository;
 import kitchenpos.product.domain.Product;
-import kitchenpos.product.repository.ProductRepository;
+import kitchenpos.product.domain.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
