+++ src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,17 +1,13 @@
 package kitchenpos.product.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.math.BigDecimal;
-import kitchenpos.menu.domain.MenuQuantity;
 import kitchenpos.product.exception.InvalidProductNameException;
 import kitchenpos.product.exception.InvalidProductPriceException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.ValueSource;
 
 @DisplayName("Product 단위 테스트")
 class ProductTest {
@@ -51,20 +47,5 @@ class ProductTest {
             assertThatThrownBy(() -> new Product("치즈버거", BigDecimal.valueOf(-1)))
                 .isExactlyInstanceOf(InvalidProductPriceException.class);
         }
-    }
-
-    @DisplayName("quantity 수 만큼 Product의 가격을 곱해서 반환한다")
-    @ParameterizedTest
-    @ValueSource(longs = {1, 3, 5})
-    void multiply(long value) {
-        // given
-        Product product = new Product("치즈버거", BigDecimal.valueOf(5_000));
-
-        // when
-        MenuQuantity menuQuantity = new MenuQuantity(value);
-        ProductPrice totalPrice = product.multiplyPrice(menuQuantity);
-
-        // then
-        assertThat(totalPrice.getValue()).isEqualTo(product.getPrice().multiply(BigDecimal.valueOf(value)));
     }
 }
